+++ src/main/java/seedu/goldencompass/command/CommandRegistry.java
@@ -6,7 +6,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class CommandRegistry {
-
     private static Map<String, Executable> executableMap = new HashMap<>();
 
     /**
