+++ src/main/java/seedu/goldencompass/parser/FlagValidator.java
@@ -1,4 +1,4 @@
-package seedu.goldencompass.preparser;
+package seedu.goldencompass.parser;
 
 import seedu.goldencompass.exception.GoldenCompassException;
 import seedu.goldencompass.set_operation.SetOperation;
