+++ src/main/java/seedu/goldencompass/parser/Preparser.java
@@ -1,4 +1,4 @@
-package seedu.goldencompass.preparser;
+package seedu.goldencompass.parser;
 
 
 
@@ -8,11 +8,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.HashMap;
 
-import static seedu.goldencompass.preparser.Config.ALL_COMMANDS;
-import static seedu.goldencompass.preparser.Config.ALL_FLAGS;
-import static seedu.goldencompass.preparser.Config.COMMAND_WORD_INDEX;
-import static seedu.goldencompass.preparser.Config.DEFAULT_FLAG;
-import static seedu.goldencompass.preparser.Config.FLAG_INDICATOR;
+import static seedu.goldencompass.parser.Config.ALL_COMMANDS;
+import static seedu.goldencompass.parser.Config.ALL_FLAGS;
+import static seedu.goldencompass.parser.Config.COMMAND_WORD_INDEX;
+import static seedu.goldencompass.parser.Config.DEFAULT_FLAG;
+import static seedu.goldencompass.parser.Config.FLAG_INDICATOR;
 
 import seedu.goldencompass.exception.GoldenCompassException;
 import seedu.goldencompass.exception.GoldenCompassParsingException;
